+++ tests/cached.rs
@@ -4,6 +4,8 @@ Full tests of macro-defined functions
 #[macro_use]
 extern crate cached;
 
+#[cfg(feature = "redis")]
+use cached::RedisCache;
 use cached::{
     proc_macro::cached, proc_macro::once, Cached, SizedCache, TimedCache, TimedSizedCache,
     UnboundCache,
@@ -1182,4 +1184,73 @@ fn test_mutable_args_once() {
     assert_eq!((2, 2), mutable_args_once(1, 1));
     assert_eq!((2, 2), mutable_args_once(1, 1));
     assert_eq!((2, 2), mutable_args_once(5, 6));
+}
+
+#[cfg(feature = "redis")]
+cached_result! {
+    UNBOUND_REDIS: RedisCache<u32, u32> = RedisCache::new();
+    fn cached_unbound_redis(n: u32) -> Result<u32, ()> = {
+        if n < 5 { Ok(n) } else { Err(()) }
+    }
+}
+
+#[cfg(feature = "redis")]
+#[test]
+fn test_cached_unbound_redis() {
+    assert!(cached_unbound_redis(2).is_ok());
+    assert!(cached_unbound_redis(4).is_ok());
+    assert!(cached_unbound_redis(6).is_err());
+    assert!(cached_unbound_redis(6).is_err());
+    assert!(cached_unbound_redis(2).is_ok());
+    assert!(cached_unbound_redis(4).is_ok());
+    {
+        let mut cache = UNBOUND_REDIS.lock().unwrap();
+        assert_eq!(2, cache.cache_size());
+        assert_eq!(2, cache.cache_hits().unwrap());
+        assert_eq!(4, cache.cache_misses().unwrap());
+        cache.cache_clear();
+    }
+}
+
+#[cfg(feature = "redis")]
+cached! {
+    TIMED_REDIS: RedisCache<u32, u32> = RedisCache::with_lifespan(2);
+    fn cached_timed_redis(n: u32) -> u32 = {
+        sleep(Duration::new(3, 0));
+        n
+    }
+}
+
+#[cfg(feature = "redis")]
+#[test]
+fn test_cached_timed_redis() {
+    cached_timed_redis(1);
+    cached_timed_redis(1);
+    {
+        let cache = TIMED_REDIS.lock().unwrap();
+        assert_eq!(1, cache.cache_misses().unwrap());
+        assert_eq!(1, cache.cache_hits().unwrap());
+    }
+    sleep(Duration::new(3, 0));
+    cached_timed_redis(1);
+    {
+        let cache = TIMED_REDIS.lock().unwrap();
+        assert_eq!(2, cache.cache_misses().unwrap());
+        assert_eq!(1, cache.cache_hits().unwrap());
+    }
+    {
+        let mut cache = TIMED_REDIS.lock().unwrap();
+        assert_eq!(2, cache.cache_set_lifespan(1).unwrap());
+        cache.cache_clear();
+    }
+    cached_timed_redis(1);
+    cached_timed_redis(1);
+    sleep(Duration::new(1, 0));
+    cached_timed_redis(1);
+    {
+        let mut cache = TIMED_REDIS.lock().unwrap();
+        assert_eq!(4, cache.cache_misses().unwrap());
+        assert_eq!(2, cache.cache_hits().unwrap());
+        cache.cache_clear();
+    }
 }